pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-environ>=0.11.2",
    "django>=5.1.4",
    "django-vite>=3.0.5",
    "inertia-django>=0.6.0",
    "psycopg>=3.2.3",
    "psycopg-binary>=3.2.3",
    "whitenoise>=6.8.2",
    "gunicorn>=23.0.0",
    "langchain>=1.0.7",
    "langchain-openai>=1.0.3",
    "langchain-community>=0.4.1",
    "pgvector>=0.4.1",
    "django-ninja>=1.5.0",
]

[dependency-groups]
dev = [
    "djlint>=1.36.4",
    "ruff>=0.14.5",
    "taskipy>=1.14.1",
]


[tool.taskipy.tasks]
# Django
run = "python manage.py runserver"
makemigrations = "python manage.py makemigrations"
migrate = "python manage.py migrate"
collectstatic = "python manage.py collectstatic --noinput -c"
createsuperuser = "python manage.py createsuperuser"

# outros
lint = "ruff check . && djlint . --quiet"
format = "ruff check . --fix && ruff format . && djlint . --reformat --quiet"

[tool.ruff]
line-length = 79
extend-exclude = ["migrations", "postgres_data", "frontend"]


[tool.ruff.lint]
preview = true
select = ["A", "B", "C", "E", "F", "I", "PL", "PT", "Q", "S", "T", "W", "DJ"]
ignore = ["E501", "PLR6301"]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.format]
preview = true
quote-style = "single"

[tool.djlint]
max_blank_lines = 1
indent = 4
tabs = true
format_css = true
format_js = true
use_gitignore = true
space_around_blocks = true
blank_line_before_tag = "include,block,endblock,crispy,if,else,endif,for,endfor,comment,endcomment,with,endwith"
blank_line_after_tag = "include,block,endblock,crispy,if,else,endif,for,endfor,comment,endcomment,with,endwith,load,extends"
ignore = "H005,H006,H007,H008,H009,H010,H011,H012,H013,H014,H015,H016,H017,H018,H019,H020,H021,H022,H023,H024,H025,H026,H027,H028,H029,H030,H031,H033,H035,H036,H037,J004,J018,T001,T002,T003,T027,T028,T032,T034"
custom_html = "c-"

[tool.djlint.css]
indent_size = 4

[tool.djlint.js]
indent_size = 4
